Add Spiral Diamonds shader
