+++ src/main/java/heronarts/lx/studio/TEApp.java
@@ -243,11 +243,12 @@ public class TEApp extends LXStudio {
       lx.registry.addPattern(RainBands.class);
       lx.registry.addPattern(SimplexPosterized.class);
       lx.registry.addPattern(SpaceExplosionFX.class);
+      lx.registry.addPattern(SpiralDiamonds.class);
       lx.registry.addPattern(TEMidiFighter64DriverPattern.class);
       lx.registry.addPattern(TESparklePattern.class);
       lx.registry.addPattern(TurbulenceLines.class);
       lx.registry.addPattern(TriangleNoise.class);
-      lx.registry.addPattern(SpiralDiamonds.class);
+      lx.registry.addPattern(OldSpiralDiamonds.class);
       lx.registry.addPattern(PulsingTriangles.class);
       lx.registry.addPattern(Fire.class);
       lx.registry.addPattern(TESolidPattern.class);
